Reorganize objects and types
